=str #15250 Use BlackholeConsumer for Flow.consume()
* last operator in front of consume() was treated specially, running
  in impl.ActorConsumer instead of the ordinary processors
